Move load_api_results.py to the API folder.
