+++ detection/detection_training/copy_checkpoints.py
