+++ detection/detection_training/experiments/eMammal_animal_human_detector/detector_label_map.pbtxt
