+++ detection/detection_training/experiments/eMammal_animal_human_detector/ssdlite_mobilenetv2_coco_weights/ssdlite_mobilenetv2_coco_weights.config
