+++ detection/detection_training/experiments/megadetector_v3/0425/pipeline.config
