+++ detection/detection_training/experiments/megadetector_v3/label_map.pbtxt
